+++ src/contracts/index.ts
@@ -23,4 +23,3 @@ export * from './solution-explorer';
 export * from './navigation/index';
 export * from './diagram/index';
 export * from './constants';
-export * from './tutorial/index';
