--- src/contracts/tutorial/Chapter.ts
@@ -1,5 +1,0 @@
-export type Chapter = {
-  name: string;
-  index: number;
-  start: Function;
-};
