--- src/contracts/tutorial/index.ts
@@ -1,1 +1,0 @@
-export * from './Chapter';
